refactor: rename to `TimeIndicator`
