+++ src/routes/[id]/components/time-indicator.svelte
